+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Next Release
 * Added support for automatically retrying throttled requests with exponential backoff to all service clients
 * Added a new config option (`gc_operation_delay`) to the DynamoDB Session Handler to specify a delay between requests
   to the service during garbage collection in order to help regulate the consumption of throughput
+* Added support for using the `us-west-2` region to the Amazon Redshift client
 * [Docs] Added a page about Waiters to the user guide
 * [Docs] Added a page about the DynamoDB Session Handler to the user guide